고객지원 writeok 파일 추가

+++ nngame/src/com/nngame/support/SupportDeleteAction.java
@@ -7,6 +7,7 @@ import javax.servlet.http.HttpSession;
 import com.nngame.action.Action;
 import com.nngame.action.ActionForward;
 import com.nngame.support.dao.SupportDAO;
+import com.nngame.user.dao.UserDTO;
 
 // 글 삭제 컨트롤러
 public class SupportDeleteAction implements Action {
@@ -14,18 +15,27 @@ public class SupportDeleteAction implements Action {
 	@Override
 	public ActionForward execute(HttpServletRequest request, HttpServletResponse response) {
 		ActionForward forward = new ActionForward();
+		HttpSession session = request.getSession();
 		
-		SupportDAO sdao = new SupportDAO();
+		// url 접근 막기 위해 user 가져옴
+		UserDTO udto = (UserDTO) session.getAttribute("udto");
 		
-		String support_num = request.getParameter("support_num");
-		
-		// 글 삭제
-		// 세션에 있는 유저 넘버 가져와서 비교후 현재 usernum가 랑 다르면 함수 실행 안되게 하기(url접근 방지)
-		sdao.deleteBoard(support_num);
-		
-		// 삭제하면 정보를 줄 필요가 없으므로 redirect로 넘겨줌
-		forward.setRedirect(true);
-		forward.setPath("/support/List");
+		if(udto == null) {
+			forward.setRedirect(true);
+			forward.setPath("/warning/approach");
+		} else {			
+			SupportDAO sdao = new SupportDAO();
+			
+			String support_num = request.getParameter("support_num");
+			
+			// 글 삭제
+			// 세션에 있는 유저 넘버 가져와서 비교후 현재 usernum가 랑 다르면 함수 실행 안되게 하기(url접근 방지)
+			sdao.deleteBoard(support_num);
+			
+			// 삭제하면 정보를 줄 필요가 없으므로 redirect로 넘겨줌
+			forward.setRedirect(true);
+			forward.setPath("/support/List");
+		}
 		
 		return forward;
 	}
